Добавил простенький сайт на шаблонах jinja2

+++ schemas/searching.py
@@ -18,7 +18,7 @@ ProductsMap = {SortingProducts.price: Product.price,
 
 class SearchingProduct(BaseModel):
     name: str = Field(..., example="Keyboard")
-    price: float = Field(..., ge=1, example=10500)
+    price: int = Field(..., ge=1, example=10500)
     rating: float = Field(..., ge=0, le=5, example=4.6)
     image: str | None = Field(example="204.jpg")
